refactor: reorganize project structure and standardize package naming
- Renamed packages 'models', 'utils', 'mappers' to singular form ('model', 'util', 'mapper') for better Java convention alignment.
- Migrated core domain models and utility classes to the new package structure.
- Removed deprecated FieldImageSaver and application.properties.
- Standardized project layout as a preparation for new feature implementation.
