+++ src/main/java/com/viking/field_passport_generator/mapper/PassportMapper.java
@@ -1,17 +1,14 @@
-package com.viking.field_passport_generator.mappers;
+package com.viking.field_passport_generator.mapper;
 
 import java.util.Collections;
 import java.util.List;
 
+import com.viking.field_passport_generator.model.*;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import com.viking.field_passport_generator.data.dto.RawFieldData;
-import com.viking.field_passport_generator.models.CropRotation;
-import com.viking.field_passport_generator.models.FieldPassport;
-import com.viking.field_passport_generator.models.GeneralInfo;
-import com.viking.field_passport_generator.models.OperationTableRow;
-import com.viking.field_passport_generator.utils.YearUtils;
+import com.viking.field_passport_generator.util.YearUtils;
 
 public class PassportMapper {
     private static final Logger log = LoggerFactory.getLogger(PassportMapper.class);
@@ -19,10 +16,12 @@ public class PassportMapper {
     /**
      * Превращает сырые данные поля и подготовленный список операций в доменный объект паспорта.
      */
-    public static FieldPassport mapToDomain(RawFieldData raw, List<OperationTableRow> operations) {
+    public static FieldPassport mapToDomain(
+            RawFieldData raw,
+            List<OperationTableRow> operations,
+            NoteSection noteSection) {
         String rawCrop = raw.crop();
-        
-        // Используем вынесенную утилиту для извлечения года
+
         String yearStr = YearUtils.extractYear(rawCrop);
         int year = yearStr.isEmpty() ? -1 : Integer.parseInt(yearStr);
 
@@ -42,6 +41,10 @@ public class PassportMapper {
             )
         );
 
-        return new FieldPassport(info, operations != null ? operations : Collections.emptyList());
+        return new FieldPassport(
+                info,
+                operations != null ? operations : Collections.emptyList(),
+                noteSection != null ? noteSection : new NoteSection(Collections.emptyList(), Collections.emptyList())
+        );
     }
 }
