+++ src/main/java/com/viking/field_passport_generator/model/FieldPassport.java
@@ -1,0 +1,9 @@
+package com.viking.field_passport_generator.model;
+
+import java.util.List;
+
+public record FieldPassport(
+    GeneralInfo generalInfo,
+    List<OperationTableRow> operations,
+    NoteSection notesSection
+) {}
