--- src/main/java/com/viking/field_passport_generator/models/FieldPassport.java
@@ -1,8 +1,0 @@
-package com.viking.field_passport_generator.models;
-
-import java.util.List;
-
-public record FieldPassport(
-    GeneralInfo generalInfo,
-    List<OperationTableRow> operations
-) {}
