+++ src/main/java/com/viking/field_passport_generator/util/JsonDataParser.java
@@ -1,8 +1,10 @@
-package com.viking.field_passport_generator.utils;
+package com.viking.field_passport_generator.util;
 
+import com.fasterxml.jackson.core.JsonParser;
 import com.fasterxml.jackson.databind.DeserializationFeature;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.json.JsonMapper;
+import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule;
 
 import java.io.InputStream;
 import java.util.Objects;
@@ -13,6 +15,8 @@ public class JsonDataParser {
     public JsonDataParser() {
         this.mapper = JsonMapper.builder()
                 .configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
+                .configure(JsonParser.Feature.AUTO_CLOSE_SOURCE, false)
+                .addModule(new JavaTimeModule())
                 .build();
     }
